LPS-109844 Move utils to api package
