+++ modules/apps/headless/headless-delivery/headless-delivery-api/src/main/resources/com/liferay/headless/delivery/util/v1_0/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
