+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/dto/v1_0/util/OpenGraphSettingsUtil.java
@@ -17,6 +17,7 @@ package com.liferay.headless.delivery.internal.dto.v1_0.util;
 import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.headless.delivery.dto.v1_0.OpenGraphSettings;
+import com.liferay.headless.delivery.dto.v1_0.util.ContentDocumentUtil;
 import com.liferay.layout.seo.model.LayoutSEOEntry;
 import com.liferay.layout.seo.service.LayoutSEOEntryLocalService;
 import com.liferay.portal.kernel.model.Layout;
